BE/etrips/requirements.txt
asgiref==3.8.1
bcrypt==4.1.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.14.4
charset-normalizer==3.3.2
click==7.1.2
cloudinary==1.40.0
colorama==0.4.6
cryptography==42.0.7
Django==4.2.13
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-oauth-toolkit==2.3.0
django-restframework==0.0.1
djangorestframework==3.15.1
drf-yasg==1.21.7
greenlet==3.0.1
idna==3.7
importlib-metadata==6.8.0
inflection==0.5.1
itsdangerous==1.1.0
Jinja2==2.11.2
jwcrypto==1.5.6
login==0.0.6
MarkupSafe==1.1.1
mysqlclient==2.2.4
oauth2-provider==0.0
oauthlib==3.2.2
packaging==24.0
pycparser==2.20
PyMySQL==0.10.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.15.0
sqlparse==0.5.0
typing-extensions==4.8.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.17.0
